prompt_chaining_lg/pyproject.toml
[project]
name = "prompt-chaining-lg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "suhaibshaikh03", email = "suhaibshaikh2003@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "langgraph>=0.2.70",
    "litellm>=1.61.0",
    "python-dotenv>=1.0.1",
]

[project.scripts]
prompt-chaining-lg = "prompt_chaining_lg:main"
piaic = "prompt_chaining_lg.main:kickoff"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
